unexport and re-locate PruneProject and helpers
