+++ txn_writer.go
@@ -11,8 +11,6 @@ import (
 	"log"
 	"os"
 	"path/filepath"
-	"sort"
-	"strings"
 
 	"github.com/golang/dep/internal/fs"
 	"github.com/golang/dep/internal/gps"
@@ -457,132 +455,9 @@ func (sw *SafeWriter) PrintPreparedActions(output *log.Logger) error {
 	return nil
 }
 
-// PruneProject removes unused packages from a project.
-func PruneProject(p *Project, sm gps.SourceManager, logger *log.Logger) error {
-	td, err := ioutil.TempDir(os.TempDir(), "dep")
-	if err != nil {
-		return errors.Wrap(err, "error while creating temp dir for writing manifest/lock/vendor")
-	}
-	defer os.RemoveAll(td)
-
-	if err := gps.WriteDepTree(td, p.Lock, sm, true); err != nil {
-		return err
-	}
-
-	var toKeep []string
-	for _, project := range p.Lock.Projects() {
-		projectRoot := string(project.Ident().ProjectRoot)
-		for _, pkg := range project.Packages() {
-			toKeep = append(toKeep, filepath.Join(projectRoot, pkg))
-		}
-	}
-
-	toDelete, err := calculatePrune(td, toKeep, logger)
-	if err != nil {
-		return err
-	}
-
-	if logger != nil {
-		if len(toDelete) > 0 {
-			logger.Println("Calculated the following directories to prune:")
-			for _, d := range toDelete {
-				logger.Printf("  %s\n", d)
-			}
-		} else {
-			logger.Println("No directories found to prune")
-		}
-	}
-
-	if err := deleteDirs(toDelete); err != nil {
-		return err
-	}
-
-	vpath := filepath.Join(p.AbsRoot, "vendor")
-	vendorbak := vpath + ".orig"
-	var failerr error
-	if _, err := os.Stat(vpath); err == nil {
-		// Move out the old vendor dir. just do it into an adjacent dir, to
-		// try to mitigate the possibility of a pointless cross-filesystem
-		// move with a temp directory.
-		if _, err := os.Stat(vendorbak); err == nil {
-			// If the adjacent dir already exists, bite the bullet and move
-			// to a proper tempdir.
-			vendorbak = filepath.Join(td, "vendor.orig")
-		}
-		failerr = fs.RenameWithFallback(vpath, vendorbak)
-		if failerr != nil {
-			goto fail
-		}
-	}
-
-	// Move in the new one.
-	failerr = fs.RenameWithFallback(td, vpath)
-	if failerr != nil {
-		goto fail
-	}
-
-	os.RemoveAll(vendorbak)
-
-	return nil
-
-fail:
-	fs.RenameWithFallback(vendorbak, vpath)
-	return failerr
-}
-
-// calculatePrune returns the paths of the packages to be deleted from vendorDir.
-func calculatePrune(vendorDir string, keep []string, logger *log.Logger) ([]string, error) {
-	if logger != nil {
-		logger.Println("Calculating prune. Checking the following packages:")
-	}
-	sort.Strings(keep)
-	toDelete := []string{}
-	err := filepath.Walk(vendorDir, func(path string, info os.FileInfo, err error) error {
-		if _, err := os.Lstat(path); err != nil {
-			return nil
-		}
-		if !info.IsDir() {
-			return nil
-		}
-		if path == vendorDir {
-			return nil
-		}
-
-		name := strings.TrimPrefix(path, vendorDir+string(filepath.Separator))
-		if logger != nil {
-			logger.Printf("  %s", name)
-		}
-		i := sort.Search(len(keep), func(i int) bool {
-			return name <= keep[i]
-		})
-		if i >= len(keep) || !strings.HasPrefix(keep[i], name) {
-			toDelete = append(toDelete, path)
-		}
-		return nil
-	})
-	return toDelete, err
-}
-
-func deleteDirs(toDelete []string) error {
-	// sort by length so we delete sub dirs first
-	sort.Sort(byLen(toDelete))
-	for _, path := range toDelete {
-		if err := os.RemoveAll(path); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
 // hasDotGit checks if a given path has .git file or directory in it.
 func hasDotGit(path string) bool {
 	gitfilepath := filepath.Join(path, ".git")
 	_, err := os.Stat(gitfilepath)
 	return err == nil
 }
-
-type byLen []string
-
-func (a byLen) Len() int           { return len(a) }
-func (a byLen) Swap(i, j int)      { a[i], a[j] = a[j], a[i] }
-func (a byLen) Less(i, j int) bool { return len(a[i]) > len(a[j]) }
